Make it easier to enable location features
Before, if the cookies weren't accepted, the user would have to go to
the bottom of the page, click "Enable cookies" link in order to access
the ability to set their own location and time zone (UTC offset).

Now, if the cookies are not accepted, the user can still click the
location icon (which now appears "disabled"), which will open a cookie
consent modal. This should make it easier to change one's mind about
cookies and enjoy the app.

Thanks @jaredlt for the suggestion.