Support custom datetime (#43)
This enables the user to set a custom date and time instead of having only access to data for the current instant.

Users can now select a custom datetime from a modal similar to the one for setting up a custom location and time zone. This feature requires cookies consent. Any instant can be used from 1900-01-01 to 2100-01-01. By default, without cookie consent, the current instant is still used.

Caelus uses optimised ephemerides for performance reasons. As a first implementation, every yearly ephemerides from 1900 to 2100 have been generated and stored. It works, however in the future we will probably implement a more dynamic way instead of having to store 200 small ephemerides.